requirements.txt
fastapi>=0.115.0
uvicorn
jinja2
pymongo
python-multipart
requests
python-dotenv
